refactor: explicit exports
