+++ packages/client/test/integration/GapFill.test.ts
@@ -2,7 +2,7 @@ import { StreamMessage } from 'streamr-client-protocol'
 import { wait } from 'streamr-test-utils'
 
 import { StreamrClient } from '../../src/StreamrClient'
-import { StreamrClientConfig } from '../../src'
+import { StreamrClientConfig } from '../../src/Config'
 import { Stream } from '../../src/Stream'
 import Subscriber from '../../src/subscribe/Subscriber'
 import { Subscription } from '../../src/subscribe/Subscription'
